backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
sqlalchemy==2.0.36
pandas==2.2.3
python-dotenv==1.0.1
textstat==0.7.4
nltk==3.9.1
google-genai==0.3.0
openai==1.51.2

pymysql==1.1.1
pymongo==4.8.0
